requirements.txt
joblib==1.5.1
numpy==2.4.2
pandas==3.0.1
PyYAML==6.0.3
scikit_learn==1.8.0
matplotlib==3.10.5
seaborn==0.13.2